Summary result return works.
Performance is (predictably) dire.
